+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_help.py
@@ -60,10 +60,22 @@ helps['aks create'] = """
           text: az aks create -g MyResourceGroup -n MyManagedCluster --ssh-key-value /path/to/publickey
         - name: Create a Kubernetes cluster with a specific version
           text: az aks create -g MyResourceGroup -n MyManagedCluster --kubernetes-version 1.8.1
-        - name: Create a Kubernetes cluster with a larger agent pool count
-          text: az aks create -g MyResourceGroup -n MyManagedCluster --agent-count 7
+        - name: Create a Kubernetes cluster with a larger node pool count
+          text: az aks create -g MyResourceGroup -n MyManagedCluster --node-count 7
+"""
+helps['aks delete'] = """
+    type: command
+    short-summary: Delete a managed Kubernetes cluster.
+"""
+helps['aks get-versions'] = """
+    type: command
+    short-summary: Get versions available to upgrade a managed Kubernetes cluster.
+"""
+helps['aks install-cli'] = """
+    type: command
+    short-summary: Download and install kubectl, the Kubernetes command line tool.
 """
 helps['aks wait'] = """
     type: command
-    short-summary: Wait for a managed Kubernetes cluster to reach a desired state.
+    short-summary: Wait for a Kubernetes cluster to reach a desired state.
 """