+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/tests/test_webapp_commands.py
@@ -192,7 +192,7 @@ class AppServiceLogTest(LiveScenarioTest):
         self.assertTrue(os.path.isdir(os.path.join(log_dir, 'LogFiles', 'kudu', 'trace')))
 
 
-class AppServicePlanSceanrioTest(ScenarioTest):
+class AppServicePlanScenarioTest(ScenarioTest):
     @ResourceGroupPreparer()
     def test_retain_plan(self, resource_group):
         webapp_name = 'webapp-quick'
@@ -210,19 +210,16 @@ class AppServicePlanSceanrioTest(ScenarioTest):
         plan = 'webapp-delete-plan2'
         self.cmd('appservice plan create -g {} -n {} -l westus'.format(resource_group, plan))
 
-        self.cmd('appservice plan update -g {} -n {} --sku S1'.format(resource_group, plan), checks=[
-             JMESPathCheckV2('name', plan),
-             JMESPathCheckV2('sku.tier', 'Standard'),
-             JMESPathCheckV2('sku.name', 'S1')
-        ])
+        self.cmd('appservice plan update -g {} -n {} --sku S1'.format(resource_group, plan),
+                 checks=[JMESPathCheckV2('name', plan),
+                         JMESPathCheckV2('sku.tier', 'Standard'),
+                         JMESPathCheckV2('sku.name', 'S1')])
 
         self.cmd('webapp create -g {} -n {} --plan {}'.format(resource_group, webapp_name, plan))
 
         self.cmd('webapp delete -g {} -n {}'.format(resource_group, webapp_name))
         # test empty service plan should be automatically deleted.
-        self.cmd('appservice plan list -g {}'.format(resource_group), checks=[
-            JMESPathCheckV2('length(@)', 0)
-        ])
+        self.cmd('appservice plan list -g {}'.format(resource_group), checks=[JMESPathCheckV2('length(@)', 0)])
 
 
 class WebappConfigureTest(ScenarioTest):
@@ -276,7 +273,7 @@ class WebappConfigureTest(ScenarioTest):
         self.assertEqual(s2['name'], 's2')
         self.assertEqual(s2['slotSetting'], False)
         self.assertEqual(s2['value'], 'bar')
-        self.assertEqual(set([x['name'] for x in result]), set(['s1', 's2', 's3', 'WEBSITE_NODE_DEFAULT_VERSION']))
+        self.assertEqual(set([x['name'] for x in result]), set(['s1', 's2', 's3']))
         # delete
         self.cmd('webapp config appsettings delete -g {} -n {} --setting-names s1 s2'
                  .format(resource_group, webapp_name)).assert_with_checks([
@@ -501,7 +498,7 @@ class WebappSlotScenarioTest(ResourceGroupVCRTestBase):
         # swap with prod and verify the git branch also switched
         self.cmd('webapp deployment slot swap -g {} -n {} -s {}'.format(self.resource_group, self.webapp, slot))
         result = self.cmd('webapp config appsettings list -g {} -n {} -s {}'.format(self.resource_group, self.webapp, slot))
-        self.assertEqual(set([x['name'] for x in result]), set(['s1', 'WEBSITE_NODE_DEFAULT_VERSION']))
+        self.assertEqual(set([x['name'] for x in result]), set(['s1']))
         # create a new slot by cloning from prod slot
         self.cmd('webapp config set -g {} -n {} --php-version {}'.format(self.resource_group, self.webapp, test_php_version))
         self.cmd('webapp deployment slot create -g {} -n {} --slot {} --configuration-source {}'.format(self.resource_group, self.webapp, slot2, self.webapp))
@@ -517,7 +514,7 @@ class WebappSlotScenarioTest(ResourceGroupVCRTestBase):
         # verify we can swap with non production slot
         self.cmd('webapp deployment slot swap -g {} -n {} --slot {} --target-slot {}'.format(self.resource_group, self.webapp, slot, slot2), checks=NoneCheck())
         result = self.cmd('webapp config appsettings list -g {} -n {} --slot {}'.format(self.resource_group, self.webapp, slot2))
-        self.assertEqual(set([x['name'] for x in result]), set(['s1', 's4', 'WEBSITE_NODE_DEFAULT_VERSION']))
+        self.assertEqual(set([x['name'] for x in result]), set(['s1', 's4']))
         result = self.cmd('webapp config connection-string list -g {} -n {} --slot {}'.format(self.resource_group, self.webapp, slot2))
         self.assertEqual(set([x['name'] for x in result]), set(['c2']))
         result = self.cmd('webapp config appsettings list -g {} -n {} --slot {}'.format(self.resource_group, self.webapp, slot))
@@ -851,6 +848,21 @@ class WebappUpdateTest(ScenarioTest):
                      JMESPathCheckV2('name', webapp_name),
                      JMESPathCheckV2('tags.foo', 'bar'),
                      JMESPathCheckV2('clientAffinityEnabled', False)])
+
+
+class WebappZipDeployScenarioTest(LiveScenarioTest):
+    @ResourceGroupPreparer(name_prefix='cli_test_webapp_zipDeploy')
+    def test_deploy_zip(self, resource_group):
+        webapp_name = self.create_random_name('webapp-zipDeploy-test', 40)
+        plan_name = self.create_random_name('webapp-zipDeploy-plan', 40)
+        zip_file = os.path.join(TEST_DIR, 'test.zip')
+        self.cmd('appservice plan create -g {} -n {} --sku S1'.format(resource_group, plan_name))
+        self.cmd('webapp create -g {} -n {} --plan {}'.format(resource_group, webapp_name, plan_name))
+        self.cmd('webapp deployment source config-zip -g {} -n {} --src "{}"'.format(resource_group, webapp_name, zip_file)).assert_with_checks([
+            JMESPathCheckV2('status', 4),
+            JMESPathCheckV2('deployer', 'Zip-Push'),
+            JMESPathCheckV2('message', 'Created via zip push deployment'),
+            JMESPathCheckV2('complete', True)])
 
 
 if __name__ == '__main__':