+++ src/command_modules/azure-cli-batchai/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+0.1.3
++++++
+
+* Fixed documentation for job list-files and stream-file
+* Added short option for providing cluster name in job create command
+
 0.1.2
 +++++
 * minor fixes