+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+2.0.18
+++++++
+* `dns`: Add support for CAA records.
+* `traffic-manager profile update`: Fix issue where profiles with endpoints could not be updated.
+* `vnet update`: Fix issue where `--dns-servers` didn't work depending on how the VNET was created (ARM deployment).
+
 2.0.17
 ++++++
 * minor fixes